No need to clarify pod is on a node
Saying:

> pods on a node can communicate with all pods on all nodes without NAT

It's unnecessary to state the pods is on a node. Doing so raises more the questions to the reader than it answers: "Why are they clarifying that the pod that is communicating is on a node?? Of course it is? Or am I missing something?". And if indeed, the reader is missing something, that something is not addressed here.